[subset] Rebase old subsetting work
In the old world this was scattered across various modules in a single
crate, but it now lives mostly in a new crate, `subset-toy`.

squashed commits:

tldr; this compiles, but I don't have compile types in place for "making
an actual font" yet, and I also have *zero* clue whether or not GPOS is
currently compiling correctly, I just sort-of know that it is compiling,
period. Now I need to go back and fill in those two gaps.

[subset] Flesh out GPOS subsetting

With this commit I'm generally satisfied that our subsetting of GPOS is
doing a reasonably good impression of a subsetter.

[subset] Tweak examples/subset_gpos